requirements.txt
pillow==10.4.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
qrcode==8.0
